requirements.txt
langchain>=0.3.0
langchain-google-genai>=2.0
langchain-community>=0.3.0
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
python-dotenv>=1.0,<2.0
